% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicity.R
\docType{data}
\name{toxicity}
\alias{toxicity}
\title{Toxicity Data}
\format{
A matrix with 171 rows and 1203 columns
}
\usage{
data(toxicity)
}
\description{
The data was created by Gul, S., Rahim, F., Isin, S. et al. (2021) \doi{https://doi.org/10.1038/s41598-021-97962-5}, downloaded and cleaned from UCI Machine Learning Repository with \doi{10.24432/C59313}. The toxicity data consists of 171 molecules with 1203 molecule descriptors.
}
\references{
\doi{https://doi.org/10.1038/s41598-021-97962-5} Gul, S., Rahim, F., Isin, S., Yilmaz, F., Ozturk, N., Turkay, M., & Kavakli, I. H. (2021). Structure-based design and classifications of small molecules regulating the circadian rhythm period. Scientific reports, 11(1), 18510.
}
\keyword{data}
\keyword{dimensional}
\keyword{high}
